COMMIT CHANGES

+++ src/main/java/com/uptc/frw/casoestudio/repository/TrabajadorRepository.java
@@ -3,5 +3,5 @@ package com.uptc.frw.casoestudio.repository;
 import com.uptc.frw.casoestudio.models.Trabajador;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TrabajadorRepository extends JpaRepository<Trabajador, Integer> {
+public interface TrabajadorRepository extends JpaRepository<Trabajador, Long> {
 }
